ci(platform): enforce uipath.platform layering with import-linter
Add an import-linter layers contract over uipath.platform, enforced as a
CI step in the lint-uipath-platform job. The contract statically forbids
import cycles and upward imports between domain packages — the class of
bug behind the 0.1.89 identity/common incident (#1787). Type-checking
imports are excluded; the contract is exhaustive so new subpackages must
be assigned a layer explicitly.

The contract is intentionally committed without ignores: lint-imports
fails at this commit on the pre-existing upward imports from
common/interrupt_models.py, fixed in the next commit.
