+++ packages/uipath-platform/src/uipath/platform/guardrails/decorators/validators/llm_as_judge.py
@@ -3,7 +3,7 @@
 from typing import Any, Sequence
 from uuid import uuid4
 
-from uipath.platform.guardrails.guardrails import (
+from uipath.platform._guardrails_service import (
     BuiltInValidatorGuardrail,
     EnumParameterValue,
     NumberParameterValue,
