+++ packages/uipath-platform/src/uipath/platform/resume_triggers/__init__.py
@@ -6,6 +6,34 @@ from ._protocol import (
     UiPathResumeTriggerHandler,
     UiPathResumeTriggerReader,
 )
+from .interrupt_models import (
+    CreateBatchTransform,
+    CreateDeepRag,
+    CreateDeepRagRaw,
+    CreateEphemeralIndex,
+    CreateEphemeralIndexRaw,
+    CreateEscalation,
+    CreateTask,
+    DocumentExtraction,
+    DocumentExtractionValidation,
+    InvokeProcess,
+    InvokeProcessRaw,
+    InvokeSystemAgent,
+    WaitBatchTransform,
+    WaitDeepRag,
+    WaitDeepRagRaw,
+    WaitDocumentExtraction,
+    WaitDocumentExtractionValidation,
+    WaitEphemeralIndex,
+    WaitEphemeralIndexRaw,
+    WaitEscalation,
+    WaitIntegrationEvent,
+    WaitJob,
+    WaitJobRaw,
+    WaitSystemAgent,
+    WaitTask,
+    WaitUntil,
+)
 
 __all__ = [
     "UiPathResumeTriggerReader",
@@ -14,4 +42,30 @@ __all__ = [
     "PropertyName",
     "TriggerMarker",
     "is_no_content_marker",
+    "CreateBatchTransform",
+    "CreateDeepRag",
+    "CreateDeepRagRaw",
+    "CreateEphemeralIndex",
+    "CreateEphemeralIndexRaw",
+    "CreateEscalation",
+    "CreateTask",
+    "DocumentExtraction",
+    "DocumentExtractionValidation",
+    "InvokeProcess",
+    "InvokeProcessRaw",
+    "InvokeSystemAgent",
+    "WaitBatchTransform",
+    "WaitDeepRag",
+    "WaitDeepRagRaw",
+    "WaitDocumentExtraction",
+    "WaitDocumentExtractionValidation",
+    "WaitEphemeralIndex",
+    "WaitEphemeralIndexRaw",
+    "WaitEscalation",
+    "WaitIntegrationEvent",
+    "WaitJob",
+    "WaitJobRaw",
+    "WaitSystemAgent",
+    "WaitTask",
+    "WaitUntil",
 ]
